refactor: restructure facades
