+++ app/facades/trust_registry.py
