--- app/routers/.gitattributes
@@ -1,13 +1,0 @@
-# Set the default behavior, in case people don't have core.autocrlf set.
-* text=auto
-
-# Declare files that will always have LF line endings on checkout.
-*.sh text eol=lf
-*.md text eol=lf
-*.json text eol=lf
-*.yaml text eol=lf
-*.yml text eol=lf
-*.conf text eol=lf
-.env* text eol=lf
-*.env text eol=lf
-manage text eol=lf
