+++ junit.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="0" time="0.229" timestamp="2021-10-25T10:44:27.609292" hostname="localhost" /></testsuites>
+<?xml version="1.0" encoding="utf-8"?><testsuites><testsuite name="pytest" errors="0" failures="0" skipped="0" tests="5" time="0.082" timestamp="2021-10-25T15:48:31.391125" hostname="Macbook-Pro-Timo.local"><testcase classname="tests.test_trust_registry_facade" name="test_actor_has_schema" time="0.006" /><testcase classname="tests.test_trust_registry_facade" name="test_actor_has_role" time="0.007" /><testcase classname="tests.test_trust_registry_facade" name="test_actor_with_role" time="0.007" /><testcase classname="tests.test_trust_registry_facade" name="test_registry_has_schema" time="0.003" /><testcase classname="tests.test_trust_registry_facade" name="test_get_did_for_actor" time="0.004" /></testsuite></testsuites>
